% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcd.R
\name{abcd.mod.year.model}
\alias{abcd.mod.year.model}
\title{abcd.mod.year.model}
\usage{
abcd.mod.year.model(par.model = c(0, 0, 0, 0, 0), Prec, PEV, Vp, ...)
}
\arguments{
\item{par.model}{A numeric vector of length 5 with the values of the abcd parameters and the initial groundwater storage}

\item{Prec}{A numeric vector with the values of the Precipitation (on a yearly basis)}

\item{PEV}{A numeric vector with the values of potential evapotranspiration}

\item{Vp}{A numeric vector with the values of the volumes of water extracted by pumping}

\item{...}{Additional parameters (used for compatibility reasons)}
}
\value{
A list with the following elements:
\itemize{
\item Qt: Numeric vector with the yearly discharges
\item SRt: Numeric vector with the surface runnoff
\item ETt: Numeric vector with the actual evapotranspiration
\item Dt: Numeric vector with the deep percolation
\item BFt: Numeric vector with the base flow
\item GFt: Numeric vector with the groundwater flow in the aquifer
\item GSt: Numeric vector with the groundwater storage
}
}
\description{
Function to calculate the components of the hydrological cycle using the modified abcd model
on a yearly basis. The modification of the original abcd model takes into account the pumping
rates and the depth to the aquifer used to calculate the variations in the groundwater levels
}
\seealso{
Other abcd_model functions: \code{\link{abcd.mod.month.model}},
  \code{\link{abcd.month.model}},
  \code{\link{abcd.year.model}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{abcd_model functions}
