% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcd.R
\name{abcd.month.model}
\alias{abcd.month.model}
\title{abcd.month.model}
\usage{
abcd.month.model(par.model = c(0, 0, 0, 0, 0, 0, 0), Prec, PEV, ...)
}
\arguments{
\item{par.model}{A numeric vector of length 5 with the values of the abcd parameters and the initial groundwater storage}

\item{Prec}{A numeric vector with the values of the Precipitation (on a yearly basis)}

\item{PEV}{A numeric vector with the values of potential evapotranspiration}

\item{...}{Additional parameters (used for compatibility reasons)}
}
\value{
A list with the following elements:
\itemize{
\item Qt: Numeric vector with the yearly discharges
\item SRt: Numeric vector with the surface runnoff
\item Wt: Numeric vector witht the available soil water
\item SMt: Numeric vector with the soil moisture
\item ETt: Numeric vector with the actual evapotranspiration
\item Dt: Numeric vector with the deep percolation
\item BFt: Numeric vector with the base flow
\item GFt: Numeric vector with the groundwater flow in the aquifer
\item GSt: Numeric vector with the groundwater storage
}
}
\description{
Function to calculate the components of the hydrological cycle using the abcd model on a monthly basis
}
\seealso{
Other abcd_model functions: \code{\link{abcd.mod.month.model}},
  \code{\link{abcd.mod.year.model}},
  \code{\link{abcd.year.model}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
