% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\name{baseflow_graphical}
\alias{baseflow_graphical}
\title{baseflow_graphical}
\usage{
baseflow_graphical(Q)
}
\arguments{
\item{Q}{A numeric vector with the discharge time series}
}
\value{
A list with the following entries:
\itemize{
\item t: numeric vector with the time
\item Qbf: numeric vector with the calculated baseflow
}
}
\description{
Function to calculate the baseflow from a discharge time series using the graphical approach
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{graphical functions}
