% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{bois_test}
\alias{bois_test}
\title{bois_test}
\usage{
bois_test(Serie1, Serie2, alpha)
}
\arguments{
\item{Serie1}{A numeric vector with the values of the hydrologic time series to be checked}

\item{Serie2}{A numeric vector with the values of the hydrologic time series to be checked}

\item{alpha}{Significance level of the test}
}
\value{
A list with the following entries:
\itemize{
\item residuals:
\item Ellipse.sup:
\item Ellipse.inf:
\item ylim:
}
}
\description{
This is a non-parametric test designed to check for the existence of a trend in the time series
of a single station or two stations using the cumulative residuals.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
