% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{calculate_diagram_moments}
\alias{calculate_diagram_moments}
\title{calculate_diagram_moments}
\usage{
calculate_diagram_moments()
}
\value{
A list with the following entries
\itemize{
\item N: coordinates of the normal distribution
\item LN3: coordinates of the three parameter Lognormal distribution
\item P3: coordinates of the Pearson distribution
\item GEV: coordinates of the Generalized Extreme Value distribution
}
}
\description{
Function to calculate the curves of the relationship between the Coefficient of Skewness and Kurtosis.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
