% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate}
\alias{calibrate}
\title{calibrate}
\usage{
calibrate(measured, fn.model, obj.fn = c("rss", "mnad", "mxad"),
  opt.method = c("lbfgs", "ga", "sa"), args)
}
\arguments{
\item{measured}{A numeric vector with measured values}

\item{fn.model}{A character string with the name of the model function}

\item{obj.fn}{A character string with the name of the objective function}

\item{opt.method}{A character string with the name of the optimization method}

\item{args}{A list with a numeric vector with the precipitation values and a numeric
vector with the PEV values}
}
\description{
Function to calibrate a water balance model using different optimization methods
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
