% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{chapman_filter}
\alias{chapman_filter}
\title{chapman_filter}
\source{
Chapman, T. G. 1991. Comment on "Evaluation of automated techniques for base flow
and recession analyses" by R. J. Nathan and T. A. McMahon, Water Resources Research, 27,
1783-1784
}
\usage{
chapman_filter(Q, a)
}
\arguments{
\item{Q}{A numeric vector with the discharge time series}

\item{a}{A numeric value specifying the filter parameter}
}
\value{
A numeric vector with the calculated baseflow
}
\description{
Function to calculate the baseflow from a discharge time series using the digital filter
proposed by Chapman1990.
}
\seealso{
Other digital_filter functions: \code{\link{eckhardt_filter}},
  \code{\link{nathan_filter}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{digital_filter functions}
