% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity.R
\name{cumulative_deviation_test}
\alias{cumulative_deviation_test}
\title{cumulative_deviation_test}
\usage{
cumulative_deviation_test(Series, prob)
}
\arguments{
\item{Series}{A numeric vector with the hydrologic time series to be tested}

\item{prob}{The significance value to be used in the test}
}
\value{
This function returns a list with the following entries:
\itemize{
\item Sk2: Rescaled adjusted partial sum 
\item Q: Change of level statistic
\item Qcrit: Critical value of Q
\item R: Range statistic
\item Rcrit: Critical value of R
\item msg: message
}
}
\description{
Homogeneity test based on the adjusted partial sums or cumulative deviations from the mean following Buishand (1982).
}
\references{
Buishand, T. SOME METHODS FOR TESTING THE HOMOGENEITY OF RAINFALL RECORDS. Journal of 
Hydrology, 1982, 1, 11-27.
}
