% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{dgumbel}
\alias{dgumbel}
\title{dgumbel}
\usage{
dgumbel(x, mu, s)
}
\arguments{
\item{x}{A numeric vector with the values of the RV}

\item{mu}{The scale factor}

\item{s}{The shape factor}
}
\value{
A numeric vector with the density values of a Gumbel distribution.
}
\description{
Function to calculate the PDF of a Gumbel distribution
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
