% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{double_mass_curve}
\alias{double_mass_curve}
\title{double_mass_curve}
\usage{
double_mass_curve(Sref, S1)
}
\arguments{
\item{Sref}{Numeric vector with the values of the reference time series}

\item{S1}{Numeric vector with the values of the hydrological time series to be checked}
}
\value{
A list with the following entries
\itemize{
\item
}
}
\description{
Function to check the consistency between two hydrological time series using the double mass curve approach.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
