% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{eckhardt_filter}
\alias{eckhardt_filter}
\title{eckhardt_filter}
\source{
Eckhardt, K.,  2005. How to construct recursive digital filters for baseflow
separation, Hydrological Processes, 19, 507-515
}
\usage{
eckhardt_filter(Q, a, BFImax)
}
\arguments{
\item{Q}{A numeric vector with the discharge time series}

\item{a}{A numeric value specifying the filter parameter}

\item{BFImax}{A numeric value specifying the maximum Base Flow Index used in the separation.}
}
\value{
A numeric vector with the calculated baseflow
}
\description{
Function to calculate the baseflow from a discharge time series using the digital filter
proposed by Eckhardt2005.
}
\seealso{
Other digital_filter functions: \code{\link{chapman_filter}},
  \code{\link{nathan_filter}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{digital_filter functions}
