% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{empirical_frequency}
\alias{empirical_frequency}
\title{empirical_frequency}
\usage{
empirical_frequency(Var1, model)
}
\arguments{
\item{Var1}{A numeric vector with the values of the hydrologic variable to be analized}

\item{model}{A character string specifying the model used to calculate the empirical frequency}
}
\value{
A list with the following entries
\itemize{
\item Var: A numeric vector with the sorted values of the hydrologic variable
\item Freq: A numeric vector with the cumulative frequency associated with each value of the
hydrologic variable
\item Treturn: Return period associated with each value of the hydrological variable calculated
using the specified model
}
}
\description{
Function to calculate the empirical frequency associated with a given dataset using a specific formula.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
