% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{estimate_parameters_mle}
\alias{estimate_parameters_mle}
\title{estimate_parameters_mle}
\usage{
estimate_parameters_mle(x, pdf.model = "normal")
}
\arguments{
\item{x}{A numeric vector with the RV values}

\item{pdf.model}{A character string with the name of the PDF to be used}
}
\value{
This function returns a list with the values of the estimated parameter for the PDF specified,
}
\description{
Function to estimate the parameters of a given distribution using Maximum Likelihood
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
