% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_budget.R
\name{evt_thornthwaite}
\alias{evt_thornthwaite}
\title{evt_thornthwaite}
\usage{
evt_thornthwaite(Temp)
}
\arguments{
\item{Temp}{A numeric vector with the mean monthly temperature}
}
\value{
A numeric vector with the calculate EVT
}
\description{
Function to calculate the potential evapotranspiration using the Thornthwaite's approach
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
