% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filling.R
\name{filling_average_station}
\alias{filling_average_station}
\title{filling_average_station}
\usage{
filling_average_station(Serie)
}
\arguments{
\item{Serie}{A numeric vector with the hydrologic time series}
}
\value{
A numeric vector with the missing values filled with the station average
}
\description{
Function to fill the missing data of a hydrologic record using the average of the station
}
\seealso{
Other fill_data functions: \code{\link{filling_month_station}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
