% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filling.R
\name{filling_month_station}
\alias{filling_month_station}
\title{filling_month_station}
\usage{
filling_month_station(Serie)
}
\arguments{
\item{Serie}{A numeric vector with the hydrologic time series to be filled}
}
\value{
A numeric vector with the corrected hydrologic time series
}
\description{
Function to fill the missing data of a hydrologic record using the monthly average
}
\seealso{
Other fill_data functions: \code{\link{filling_average_station}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
