% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity.R
\name{locate}
\alias{locate}
\title{locate}
\usage{
locate(vector_, value)
}
\arguments{
\item{vector_}{A numeric vector}

\item{value}{a numeric value}
}
\value{
This function returns the first position for which the value occurs in a vector
}
\description{
Function to locate a specific value inside a vector
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
