% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{log_nash_sutcliffe}
\alias{log_nash_sutcliffe}
\title{log_nash_sutcliffe}
\usage{
log_nash_sutcliffe(measured, calculated, ...)
}
\arguments{
\item{measured}{A numeric vector with the measured values}

\item{calculated}{A numeric vector with the calculated values}

\item{...}{Additional parameters}
}
\value{
A numeric value with the logarithmic Nash-Sutcliffe criteria.
}
\description{
Function to calculates the logarithmic Nash-Sutcliffe criteria between a vector with measured values and
other vector with calculated values using a mathematical model. This criteria applies only
to dicharge data.
}
\seealso{
Other calibration functions: \code{\link{mass_balance_error}},
  \code{\link{max_absolute_deviation}},
  \code{\link{mean_absolute_deviation}},
  \code{\link{nash_sutcliffe}}, \code{\link{objective.fn}},
  \code{\link{schultz_criteria}},
  \code{\link{sum_squared_residuals}},
  \code{\link{uncertainty_quantification_boot}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{calibration functions}
