% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{mean_absolute_deviation}
\alias{mean_absolute_deviation}
\title{mean_absolute_deviation}
\usage{
mean_absolute_deviation(measured, calculated, ...)
}
\arguments{
\item{measured}{A numeric vector with the measured values}

\item{calculated}{A numeric vector with the calculated values}

\item{...}{Additional parameters}
}
\value{
A numeric value with the mean absolute deviation.
}
\description{
Function to calculates the mean absolute deviation between a vector with measured values and
other vector with calculated values using a mathematical model.
}
\seealso{
Other calibration functions: \code{\link{log_nash_sutcliffe}},
  \code{\link{mass_balance_error}},
  \code{\link{max_absolute_deviation}},
  \code{\link{nash_sutcliffe}}, \code{\link{objective.fn}},
  \code{\link{schultz_criteria}},
  \code{\link{sum_squared_residuals}},
  \code{\link{uncertainty_quantification_boot}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
