% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{nathan_filter}
\alias{nathan_filter}
\title{nathan_filter}
\source{
Nathan, R. and McMahon, T. 1991. Evaluation of automated techniques for base flow
and recession analyses, Water Resources Research, 27, 1783-1784
}
\usage{
nathan_filter(Q, a)
}
\arguments{
\item{Q}{A numeric vector with the discharge time series}

\item{a}{A numeric value specifying the filter parameter}
}
\value{
A numeric vector with the calculated baseflow
}
\description{
Function to calculate the baseflow from a discharge time series using the digital filter
proposed by Nathan1991.
}
\seealso{
Other digital_filter functions: \code{\link{chapman_filter}},
  \code{\link{eckhardt_filter}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
