% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{objective.fn}
\alias{objective.fn}
\title{objective.fn}
\usage{
objective.fn(x, fn.model, params, fit.fn, measured)
}
\arguments{
\item{x}{A numeric vector with the parameters to be estimated}

\item{fn.model}{A character string with the name of the model used to evaluate the parameters}

\item{params}{A list with the arguments of the fn.model}

\item{fit.fn}{A character string with the name of the function used to calculate the difference
between the calculated and measured values}

\item{measured}{A numeric vector with the measured values}
}
\value{
A numeric value of the objective function
}
\description{
Calculates the objetive function
}
\seealso{
Other calibration functions: \code{\link{log_nash_sutcliffe}},
  \code{\link{mass_balance_error}},
  \code{\link{max_absolute_deviation}},
  \code{\link{mean_absolute_deviation}},
  \code{\link{nash_sutcliffe}},
  \code{\link{schultz_criteria}},
  \code{\link{sum_squared_residuals}},
  \code{\link{uncertainty_quantification_boot}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
