% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{pgumbel}
\alias{pgumbel}
\title{pgumbel}
\usage{
pgumbel(q, mu, s)
}
\arguments{
\item{q}{A numeric vector with quantiles of the RV}

\item{mu}{The scale factor}

\item{s}{The shape factor}
}
\value{
A numeric vector with the cumulative probability values of a Gumbel distribution.
}
\description{
Function to calculate the CDF of a Gumbel distribution
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
