% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{probability_plot}
\alias{probability_plot}
\title{probability_plot}
\usage{
probability_plot(Var, model)
}
\arguments{
\item{Var}{A numeric vector with the values of the variable to be checked}

\item{model}{A character string specifying the distribution function}
}
\value{
This function returns a list with the following entries:
\itemize{
\item Var: Measured variable
\item z: Theoretical variates
\item model: PDF model
}
}
\description{
Function to calculate the measured and theoretical quantiles required in different probability
plots.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
