% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{qgumbel}
\alias{qgumbel}
\title{qgumbel}
\usage{
qgumbel(p, mu, s)
}
\arguments{
\item{p}{A numeric vector with cumulative probability values}

\item{mu}{The scale factor}

\item{s}{The shape factor}
}
\value{
A numeric vector with the quantiles of a Gumbel distribution.
}
\description{
Function to calculate the quantiles of a Gumbel distribution
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
