% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{sum_squared_residuals}
\alias{sum_squared_residuals}
\title{sum_squared_residuals}
\usage{
sum_squared_residuals(measured, calculated, ...)
}
\arguments{
\item{measured}{A numeric vector with the measured values}

\item{calculated}{A numeric vector with the calculated values}

\item{...}{Additional parameters}
}
\value{
A numeric value with the sum of squared residuals.
}
\description{
Function to calculate the sum of squared residuals between a vector of measured values and other
with the calculated values using a model.
}
\seealso{
Other calibration functions: \code{\link{log_nash_sutcliffe}},
  \code{\link{mass_balance_error}},
  \code{\link{max_absolute_deviation}},
  \code{\link{mean_absolute_deviation}},
  \code{\link{nash_sutcliffe}}, \code{\link{objective.fn}},
  \code{\link{schultz_criteria}},
  \code{\link{uncertainty_quantification_boot}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
