% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_R_cumulative_deviation}
\alias{table_R_cumulative_deviation}
\title{Table of the critical values of the range statistic R}
\format{A data frame with 7 rows and 4 variables
\describe{
\item{n}{Number of samples}
\item{X0.90}{Critical values of the range statistic R at 90\%}
\item{X0.95}{Critical values of the range statistic R at 95\%}   
\item{X0.99}{Critical values of the range statistic R at 99\%}
}}
\source{
Buishand, T. SOME METHODS FOR TESTING THE HOMOGENEITY OF RAINFALL RECORDS. Journal of 
Hydrology, 1982, 1, 11-27.
}
\usage{
table_R_cumulative_deviation
}
\description{
Table of the critical values of the range statistic R
}
\keyword{datasets}
