% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{uncertainty_quantification_boot}
\alias{uncertainty_quantification_boot}
\title{uncertainty_quantification_boot}
\usage{
uncertainty_quantification_boot(measured, fn.model, obj.fn = c("rss", "mnad",
  "mxad"), opt.method = c("lbfgs", "ga", "sa"), args, level = c(0.025,
  0.975), neval = 100, seed = 12345)
}
\arguments{
\item{measured}{A numeric vector with measured values}

\item{fn.model}{A character string with the name of the model function}

\item{obj.fn}{A character string with the name of the objective function}

\item{opt.method}{A character string with the name of the optimization method}

\item{args}{A list with a numeric vector with the precipitation values and a numeric
vector with the PEV values}

\item{level}{A numeric vector with the values of the quantiles associated with the
confidence interval to estimate}

\item{neval}{An integer specifying the number of model evaluation used to estimate the
confidence intervals}

\item{seed}{A random seed}
}
\value{
A list with the following entries:
\itemize{
\item
\item
}
}
\description{
Function to calculate the confidence intervals of the model parameters using a bootstrapping
approach.
}
\seealso{
Other calibration functions: \code{\link{log_nash_sutcliffe}},
  \code{\link{mass_balance_error}},
  \code{\link{max_absolute_deviation}},
  \code{\link{mean_absolute_deviation}},
  \code{\link{nash_sutcliffe}}, \code{\link{objective.fn}},
  \code{\link{schultz_criteria}},
  \code{\link{sum_squared_residuals}}
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
