% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity.R
\name{von_neumann_test}
\alias{von_neumann_test}
\title{von_neumann_test}
\usage{
von_neumann_test(Series)
}
\arguments{
\item{Series}{A numeric vector with the hydrologic time series}
}
\value{
A list with the following entries:
\itemize{
\item NN: A numeric value with the Von Neuman statistic
\item msg: A character string with the result of the test
}
}
\description{
Von Neumann test for homogenity of time series
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
\concept{homogenity functions}
