% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_budget.R
\name{water_budget_direct}
\alias{water_budget_direct}
\title{water_budget_direct}
\usage{
water_budget_direct(Prec, EVT, Rmax)
}
\arguments{
\item{Prec}{Numeric vector with the precipitation values}

\item{EVT}{Numeric vector with the evapotranspiration values}

\item{Rmax}{Maximum soil retention (in mm)}
}
\value{
This function returns a list with the following entries:
\itemize{
\item R: soil storage
\item RV: Soil storage variation
\item ETR: Real evapotranspiration
\item WD: Water deficit
\item WE: Water excess
\item RN: Runoff
\item PC: Deep percolation
}
}
\description{
Function to calculate the water budget using the direct method
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
