% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{calculate_diagram_moments}
\alias{calculate_diagram_moments}
\title{calculate_diagram_moments}
\usage{
calculate_diagram_moments()
}
\value{
A list with the following entries
\itemize{
\item N
\item LN3
\item P3
\item GEV
}
}
\description{
Function to calculate the curves of the relationship between the Coefficient of Skewness and Kurtosis.
}
\author{
Oscar Garcia-Cabrejo \email{khaors@gmail.com}
}
